api/requirements.txt
Pillow==9.1.1 
Flask==2.1.2
Flask-Cors==3.0.10 
Werkzeug==2.1.2 
APScheduler==3.9.1 